Support of DEFAULT_ALLOC_WEI

+++ run.bash
@@ -10,6 +10,7 @@ export RPCARGS='--rpc --rpcaddr 0.0.0.0 --rpccorsdomain=* --rpcapi "admin,debug,
 export DEFAULT_PASSWORD_PATH=${DEFAULT_PASSWORD_PATH:-"/cliquebait/default-password"}
 export ACCOUNTS_TO_CREATE=${ACCOUNTS_TO_CREATE:-"5"}
 export EXTERNAL_ALLOCS=${EXTERNAL_ALLOCS:-""}
+export DEFAULT_ALLOC_WEI=${DEFAULT_ALLOC_WEI:-"1000000000000000000000"}
 
 if $VERCOMP $STRIPPED_GETH_VERSION '>=' 1.8.0; then
 	echo 'adding --rpcvhosts=* as we are in geth >= v1.8.0'
@@ -43,15 +44,22 @@ function initialize_geth() {
 	geth --datadir=$GETHDATADIR init $GETHROOT/genesis.json
 }
 
-# $1 = source genesis, $2 = the FULL address (with 0x)
+# $1 = source genesis, $2 = the FULL address (with 0x), $3 = allocation of Wei
 function give_ether_in_genesis() {
 	mkdir -p /tmp/cliquebait/give_ether_in_genesis
 	BAREADDRESS=`echo $2 | sed s/0x//`
-	MOREALLOCS="{\"$BAREADDRESS\": {\"balance\": \"0x200000000000000000000000000000000000000000000000000000000000000\"}}"
-	cat $1 | jq ".alloc += $MOREALLOCS" > /tmp/cliquebait/give_ether_in_genesis/new_genesis.json
-	rm $1
-	mv /tmp/cliquebait/give_ether_in_genesis/new_genesis.json $1
-	rm -rf /tmp/cliquebait/give_ether_in_genesis
+	CLEANED_ALLOC_WEI=$(echo $3 | sed -e 's/^0x//')
+	ALLOC_WEI=$(echo $CLEANED_ALLOC_WEI | grep -o "[0-9A-Fa-f]\{${#CLEANED_ALLOC_WEI}\}$")
+	if [ -z "$ALLOC_WEI" ]; then
+		echo $DEFAULT_ALLOC_WEI is an invalid value of Wei given by DEFAULT_ALLOC_WEI.
+	else
+		echo "matched $ALLOC_WEI Wei for allocation!"
+		MOREALLOCS="{\"$BAREADDRESS\": {\"balance\": \"0x$ALLOC_WEI\"}}"
+		cat $1 | jq ".alloc += $MOREALLOCS" > /tmp/cliquebait/give_ether_in_genesis/new_genesis.json
+		rm $1
+		mv /tmp/cliquebait/give_ether_in_genesis/new_genesis.json $1
+		rm -rf /tmp/cliquebait/give_ether_in_genesis
+	fi
 }
 
 # this works because we have DEPLOY_ACCOUNT_ADDRESS in place of the actual hex in the extraData field of the genesis JSON
@@ -120,7 +128,7 @@ function initialize_cliquebait() {
 	fi
 
 	# give all the loaded accounts tons of ether
-	for address in `cat $CBROOT/accounts`; do give_ether_in_genesis $CBROOT/genesis.json $address; done
+	for address in `cat $CBROOT/accounts`; do give_ether_in_genesis $CBROOT/genesis.json $address $DEFAULT_ALLOC_WEI; done
 
 	# give accounts listed in $EXTERNAL_ALLOCS some ether as well
 	for extra in $(echo $EXTERNAL_ALLOCS | tr ',' ' '); do
@@ -129,7 +137,7 @@ function initialize_cliquebait() {
 			echo $extra is an invalid address and will not have any ether allocated.
 		else
 			echo "matched $extra for allocation!"
-			give_ether_in_genesis $CBROOT/genesis.json 0x$cleaned_extra
+			give_ether_in_genesis $CBROOT/genesis.json 0x$cleaned_extra $DEFAULT_ALLOC_WEI
 		fi
 	done
